requirements.txt
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pandas==2.0.3
numpy==1.24.3
python-dotenv==1.0.0
pyyaml==6.0.1
jupyter==1.0.0
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3 